6.10 Raty fixes

+++ app/models/product.rb
@@ -22,7 +22,7 @@ class Product < ApplicationRecord
   end
 
   def average_rating
-    comments.average(:rating).to_f
+    comments.average(:rating).to_i
   end
 
 end
